requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-SocketIO==5.6.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyserial==3.5
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.7
wsproto==1.3.2
